+++ src/mixins/graphics/mixinEllipse.js
@@ -1,8 +1,8 @@
-import { semiMajorAxis, semiMinorAxis, height, extrudedHeight, rotation, stRotation, granularity } from '../mixinProps'
+import { semiMajorAxis, semiMinorAxis, height, extrudedHeight, rotation, granularity } from '../mixinProps'
 
 /**
  * vc-graphics-ellipse base props mixins
  */
 export default {
-  mixins: [semiMajorAxis, semiMinorAxis, height, extrudedHeight, rotation, stRotation, granularity]
+  mixins: [semiMajorAxis, semiMinorAxis, height, extrudedHeight, rotation, granularity]
 }